Revert "fix: Clarity once again blocks *Rage berserk (Mantis 12640)"
This reverts commit 51f8dfeb475ed0e39ea3046524384fdaf5c92adb.

Commit 51f8dfe intentionally made *Rage on artefacts an intentional
source of berserk, as equipping the item and attacking in melee
is intentional, and otherwise Chei players could get access to penance
free berserk (the hates item code is only for generating the warning
prompt, penance comes from actually using it).

This revert clarifies the description of *Rage to make it read as
intentional source of berserk (like the potion or +Rage), which clarity
does not block. It also displays the items contribution to the berserk chance,
and makes these contributions linear so the probability is clear to
players.