feat: remove evoked berserk, increase *Rage berserk chance
Step-by-step we have been cutting back on items that add active
abilities to the 'a' menu: even with the quiver these abilities are
often forgotten by players for ui reasons since they only come from
artefacts (+Inv is the remaing exception) so aren't frequently
encountered. A player can't forget to use a "passive" ego.

*Rage was meant as a bad property. With a 5% chance to berserk it was
small enough that one could take a swing here and there and not expect
to berserk (making it an unreliable source of zerk), but large enough
that a player trying to use it had to approach fights expecting to go
berserk in every fight.

This commit removes +Rage in favor of giving *Rage a greater chance to
trigger berserk, currently 20%. This leans into the "always be ready to
zerk", while being low enough that it's not a guaranteed immediate
berserk in desparate situations. The insrcription is changed simply to
"Rage".

When it comes to changing unrands:
- Leech and Bear Spirit might want lower (or 0) Angry values, but for
  now they both get the standard 20% for clarity;
- the necklace of bloodlust, which has a swap penalty and extra zerk
  powers, gets a 50% zerk chance
- the wrath of Trog, with its swap penalty, keeps its 50% zerk chance
- the Zealot's sword gets the standard 20% chance;

The chance contribution is shown in the item description, and the total
zerk chance is shown in the character overview if nonzero.