feat: combine berserkitis and item Rage
The separate displays and (slightly) separate semantics for "go berserk
on melee hit" confounded the issue, especially since berserkitis'
chances are so small at low levels. This commit changes berserkitis to
simply grant angry percentage points, identical to other sources of
random berserk.

It does mean that berserkitis now works through clarity and causes
Cheibriados penance. This is consistent with the previous commits:
berserkitis only triggers on melee attacks, which are intentional. If
this is too severe some grace can be hacked in for Chei followers, but
the fraction of characters that will find berserkitis much worse than
they already do is tiny.