Replace Leech's *Rage with Harm
*Rage is a fairly severe downside and means the player should expect to
berserk in every combat, which heavily limits the chances of any items with
it from being actually used. It's very thematic on the Wrath of Trog and
necklace of Bloodlust, and the zealot's sword also has a unique (although
perhaps hard-to-utilise) interaction with berserk, but the other unrands
with *Rage could use different downsides to encourage still using them.

The hat of the Bear Spirit could use a new idea, but for Leech, replace
*Rage with Harm (and drain-on-unequip to prevent it from being freely
swappable). The only other use of Harm on an unrand is the robe of
Misfortune, so Leech's guaranteed vampiricism seems like a fun place to
encourage trying to work around Harm's downsides.