Add %syn lines for granite statue variants
As recommended here: https://github.com/crawl/crawl/issues/1947
This allows specific granite statue variants to be used in vault-making.

NOTE (and help please if possible): 
I can't seem to get "%syn DNGN_STATUE_DWARF" to work, probably because it's been declared as the default/random granite statue? See:
> statue_dwarf DNGN_GRANITE_STATUE
Adding the %syn line below this doesn't work correctly. Is there a way to fix this? I'd love to be able to use the dwarf statue in future dwarf-themed vaults.