dreamdust_arrival_beginning_and_end
Arrival vault that resembles the central orb chamber in Zot:5. Thought it would be cute (especially for new players) to realize it at the end of the orb run.

QUESTION:  Is there a clever way to give different names to dungeon features (granite statues in this case)?
For example, if I use:
> : set_feature_name("granite_statue", "a sculpture of the fabled Orb of Zot")

That will also affect the description of the orb guardian statues. I'm not sure how to work around this so I can give them each a unique description.