feat: replace hell contam with temporary mutations
In the game of one of these things is not like the other, the contam
hell effect produced permanent strategic damage, the other two produced
XP gated strategic damage. Contam hits a few more things other than
mutations (putting a brake on miscasts and preventing invis) but the
interaction with the mutation system is the noticable part in hell. At
the rate we'd like to apply hell effects, however, it gets to be a bit
much in terms of permanent damage. So let's try temp mutations.