Don't crash when failing to pacify nothing
Targeting empty space with pacification is allowed so that invisible
monsters can be pacified, but doing so still costs a turn to avoid leaking
information. However if the attempt to use the ability rolled a failure, it
would instead crash when targeting an empty space due to a missing
fail_check().