dreamdust_elfheim
The home of Duvessa and Dowan, This vault might belong is some other file since it's mostly decorative? Not sure.
The more-practical Duvessa has training dummies in her room, while the vain Dowan has luxurious marble tiling in his room as well as a large mirror to admire himself with.