Axe ironbound beastmasters
Ironbound beastmasters were a whole pile of cool ideas, none of which quite
panned out.

- They depended very heavily on their bands, which tends to work poorly
  (it's pretty common for monsters to get separated)
- They had some hacks to guarantee they'd only generate with bands, which
  broke monster queries and still didn't prevent them from getting separated.
- The effects of their spell was poorly communicated to players - the energy
  system is a very secret under-the-hood effect, unlike e.g. haste, which has
  a visible effect.
- They weren't tuned to the point of being very strong. (This is the most
  fixable issue.)

There was some discussion of replacing them with a summoner that calls and
berserks harpies and such, and we could still do that. For now, though, let's
roll back this experiment.

Partially reverts 0b001bae985.