fix: allow unrand air enhancer stacking (#2134)
e08c5e17fd added an air enhancer to the amulet of the air, but unlike the
rest of the spell enhancers in the game, it didn't permit stacking this
with the elemental staff. This looks to me like an oversight rather than a
balance issue, (especially given that you could still stack the amulet with
a vanilla staff of air), so here's a fix.