fix: clean up failed crawl processes correctly in webtiles
I think this should fix the lobby in scenarios we see where something
goes very wrong with the webtiles server and the lobby ends up looking
like there are many duplicate games. What was happening is that if an
exception prevents the crawl process from starting altogether, the
cleanup code was also throwing an exception on this first logging call
because self.process was None. This prevented the inprogress file from
getting removed, and the webserver's ongoing game table from getting
cleaned up.

This will unfortunately make it *harder* to see when a webtiles server
is completely borked, so possibly something should be done about that. I
did add an error message to the player when this happens. (In the
process of adding this, I attempted to alleviate a race condition that
was preventing exit dialogs from being shown when a process never
started. Hopefully I did so correctly. Basically, multiple go_lobby
messages in close timing will wipe out an exit dialog if it's already
been shown.)