wip: rework and streamline the dreamshard necklace
- Dreams portal and branch removed
- Amulet effect changed to a large heal on evoke plus summoning a modest
number of dream spectres (which make good use of some of the excellent,
but currently unused artwork).
This keeps the strategic intention of the item whilst boiling it down
to a much more streamlined implementation.

RIP dream portal, but welcome dream spectres (incl hydrataur).

Still wip but close to next point to pause for review.