Cut bailey spawn depth
Bailey portals spawned between D:7-14 and in Orc:1. While at first glance this
looks like a small portion of the game, players get massively stronger over
those floors. D:7 characters might still be working on their starting kit,
whereas a character on D:14 or in Orc is likely well established and has
probably cleared Lair. The monsters in many existing Bailey vaults are similar
to Orc spawns as well, so they tend to be threatening in the early part of the
Bailey spawn range but mundane in the later part. Thus, Bailey will now spawn
between D:7-10 and the chance per floor is doubled to keep the overall
likelihood of seeing a bailey roughly the same.