Adjusting the power cost of Fedhas
Fedhas' plants are very powerful and the piety cost of them seems reasonable. However, in actual battles, due to restrictions on plants being unable to move, players had to summon plants again or bring their enemies to the place where they were. Also, Fedhas didn't give the player useful passive even if there were no plants, so the player had to  pay a lot of fee for every battle to get God's help.

So I'm going to lower the cost of the piety a little bit. However, the abuse of power is also not desired, so MP costs are increased in return. Instead of being able to summon plants more often, players will now have difficulty using power several times in a single battle(Lack of MP).