Rework bailey_axe_3
This was formerly one of two 'boomerang kobolds behind deep water' baileys. Now
it is a kobold brigand bailey; they replace a few of the regular kobolds
(which are no longer behind deep water) and the other weak enemies. Loot has
been updated to include a few stacks of specialty darts.