feat: new yred: remove injury mirror
Injury Mirror was the most awkward of Yred's abilities. It cost a lot of
piety if it activated, but getting a big activation meant taking a lot
of damage. In comparison, Drain Life gains hitpoints and costs a
predictible amount of piety.

It's possible there's an injurry-mirror or damage-sharing ability that
works with the re-designed Yredelemnul "zombies are piety" principle,
beyond just having allies to soak up damage. For now, this gets the axe.