Remove archer monsters hidden bonus damage
This commit removes the HD * 4/3 bonus damage from monsters with the archer tag. This bonus, besides being hidden, lacks an analogous bonus in melee monsters (those with the fighter tag). The end result of this is rather unintuitive from the player's perspective, since when the player uses ranged weapons, they usually hit for a little bit less than their melee counterparts, and this is dramatically different from monsters (a merfolk javelineer, for example, hits for a little bit more than an impaler with a +9 demon trident).