dreamdust_black_sun_cult
Vault featuring a fire- and necromancy- based cult. In the Dungeon, players will encounter some (mundane) demonspawn worshipping a sun demon, while in Zot they'll find powerful draconians and black sun demonspawn worshipping an orb of fire.
Rewards are a book (either a manual or a spellbook) and a fire-related item.