Two large themed vaults
dreamdust_isle_of_ice: An ice-themed vault on the edge of the map.
Original inspiration is how Antarctica looks on a 2D world map, but I'm
a terrible artis. So quickly gave up on replicating it.

dreamdust_labyrinth: An homage to the old labyrinths. I gave it the
lowest WEIGHT since it's a large vault and I think it works best as a
very rare easter egg.

Committer's Note: Combined commits and tweaked commit messages.