fix: Rework the border of a ghost vault
In ebering_ghost_funeral_grove, the vault border is made of trees with a
destruction veto. This isn't revealed to players except when they
attempt to destroy the tree, which is spoilery. Additionally it has odd
gameplay side effects with respect to certain spell effects. This commit
adds a border of stone to the vault, retaining the use of trees for the
vault's interior, and removes the tree destruction veto. The vault's
theme still very much comes across with this change.