feat: Simplify Vaults branch ghost vaults
Currently for ghost vaults in the Vaults branch, we maintain separate
copies of existing vaults. This was done because Vaults branch ghost
vaults must be defined as so-called vault rooms that have special
tagging and, in some cases, changes to layout relative to the original
ghost vault. However in all cases this wasn't actually necessary, since
the original vault could have relatively minor modifications allowing it
to be uses as-is in both branches.

This commit merges the Vaults branch vaults back into the original
vaults, mainly by copying over any Vaults-specific monster and loot
definitions and by generalizing the ghost vault setup code to do any
required Vaults branch tagging and setup. We remove the
`vaults_ghost_setup()` function, and modify `ghost_setup()` to have a
boolean argument that performs any setup related to the Vaults branch
when the argument is true.

We also modify the terrain of these vaults to give more spacing near
areas that could create single square corridors should the vault place
as a Vaults room. In some cases this just adds a row of floor near the
vault entrance, but other vaults need an addition border of floor to
surround the entire vault. Transporter vaults are now marked `passable`
(for Vaults) and `transparent` with relevant sections masked with
`opaque`. The transparent/opaque masking isn't required, but helps the
layout code properly determine the vault's connectivity.

The ghost vault setup function is now documented. The ghost vault
guidelines will be updated in the future.