feat: Improve a monster weapon selection function
The function `dgn.monster_weapon` is used to give vault monsters higher
quality weapons appropriate to their function. Add one-handed and
two-handed classes for "warrior" and "knight" to allow vaults using this
function to specify the handedness of the desired weapon.  This is
useful when when the original monster always recieves a weapon of a
specific handedness, e.g. tengu monsters always have a buckler and
ancient champions always use a two-hander. Update a few ghost vaults to
use these new classes where appropriate.