Break doors
In certain circumstances, it's possible for players to door-dance,
repeatedly closing doors in the face of an enemy who then opens the
door, etc. This allows the player time to heal and gives a decent
chance the enemy will simply forget about them. There's already a
chance for doors to creak, creating noise, but this has a very
limited impact on the tactic.

To weaken this somewhat (in context of recent changes to random
energy), give a 1/3 chance of doors breaking whenever monsters open
them. This turns them into 'broken doors', which can be re-locked by
vault wardens but are otherwise useless. (If we simply destroyed the
doors, players would be incentivized to repeatedly open doors until
they broke in some circumstances in Vaults, to mitigate vault wardens.)

This should have fairly little impact on most players, but it sure was
a pain in the butt to make crude sprites for.