Remove random energy
Random energy has a long and dismal history. It was invented as a
means to discourage repetitive and boring play patterns (pillar-dancing,
door-dancing, break-dancing, etc), but signally failed to do so.
Instead, it obfuscated game mechanics, added situations that felt
'unfair', and generally kludged everything up.

It was a bold experiment, but it's been clear for many years now
that it's failed. Let's move on.