Implement closing attacks
An experimental way to discourage long-distance luring or so-called
'pillar dancing', which tends to be rather repetitive and not very
fun. Gives monsters a 1/5 chance to get a free attack when the player
moves away from them. This only costs energy for enemies that attack
slowly, like ogres and juggernauts.

This should have a minimal impact on most normal fights, and even allow
repositioning across a reasonable distance with adjacent enemies.