refactor: split DNGN_ENTER_HELL into one feature per destination
Replace DNGN_ENTER_HELL with DNGN_EXIT_COCYTUS, DNGN_EXIT_DIS,
DNGN_EXIT_GEHENNA or DNGN_EXIT_TARTARUS for gates leading out of those
branches.

This means that the descriptions for these are more appropriate, and
makes it easier to change how one of them looks in a console build.

[ Committer's note: Closes #2173. Tweaked descriptions and made these
  PORTAL_EXIT features. ]