Remove references to Hell from feat_is_branch_entrance().
feat_is_hell_subbranch_exit() can simply be deleted, as the function returns false whether it's listed separately or not.
The response to DNGN_ENTER_HELL changes from false to true, so change the functions which call it to accommodate this.

The things which should behave differently are:
Hints mode now treats DNGN_ENTER_HELL as a branch entrance.
In lua (dlib), is_branch_entrance() now returns true for DNGN_ENTER_HELL. This isn't used in the Crawl source at present.