fix: ghost vault logic
The ghost vault logic changes in 17de5e62 were partially incorrect and
messed up ghost vault placement. We need to have ghost vaults placing
outside of the Vaults branch always set a CHANCE, but not set CHANCE for
the Vaults branch. Furthermore ghost vaults exclusive to the Vaults
branch should only exist as so-called Vault rooms in the file
`vaults_rooms_ghost.des`. The vaults in this file have no set DEPTH so
that they never attempt to place as minivaults but are still selectable
by the `vaults_ghost` tag. However 17de5e62 had these vaults also set a
CHANCE, which caused this to be applied to a default depth of *all
levels*, even ones not included in the ghost vault range.  Furthermore
the conditional in `ghost_setup()` incorrectly didn't set CHANCE if the
`vaults_setup` argument was true, even for ghost vaults that placed both
in Vaults and outside of Vaults.

This commit fixes these issues. For Vaults exclusive ghost vaults in
`vaults_rooms_ghost.des`, we have them set a new `set_chance` arg for
`ghost_setup()` to false to prevent setting the CHANCE at all. The
vaults in this file thus only rely on the `vaults_ghost` tag to be used
(and only in the Vaults branch). This argument defaults to true, so
ghost vaults outside of this file will get their CHANCE set properly for
non-Vaults branches. Furthermore we give any such vault a CHANCE weight
of 0 specifically in Vaults. This prevents a ghost vault that places
both outside of Vaults (as a CHANCE vault) and inside of Vaults (as a
ghost vault room) from ever independently placing in Vaults branch as a
minivault via CHANCE.