Reduce the electrocution brand's chance to trigger
Electrocution highly overperforms as a weapon brand - not only is it rare
for monsters to resist it, its damage was high enough to compete with
multiplicative brands such as vorpal/freezing/flaming even on larger,
slower weapons. Reduce its chance to trigger from 1/3 to 1/4 per hit, to
make it slightly less of an obvious best choice.