feat: Merge door vaults maps
For the variations of door_vault placing in the Vaults branch, we
previously had separate vaults for these. Similar to to the ghost vault
merge done in 17de5e6, this commit merges the Vaults branch variants
into their parent vaults, adjusting the parent to have a ring of floor
tiles and the necessary tags.