fix: condense logging on webtiles connection timeout
Based on 7bb4a8268b, one thing I can say is that in Tornado 4+, when a
timeout finally reaches the point of closing out the socket, there can
be a really large number of async write Futures waiting to be failed,
that all get their callbacks triggered within a few ms of each other.
This situation in and of itself indicates some issues in our code, e.g.
one culprit apparently is lobby messages, which look to be really
inefficient and scale badly with the number of players on the server.
However, it seems possible here that the server thrashing conditions are
triggered by just too much logging in response to this sudden burst of
callbacks. I still haven't replicated any of this locally, so this is
all somewhat speculative. However, for exceptions indicating that the
connection is closed, now that I have seen some logging, I don't think
there's a huge need to keep the verbose logging on (it can always be
restored if I need more data).

This commit logs just the first one, and keeps track of the total number
of failed callbacks on connection end, printing that in the final close
message.  Regular exceptions will still be fully logged, if they happen.
This commit also adds correct handling for StreamClosedError in tornado
4.