feat: Roomify Vaults branch Desolation entries
Similar to what was done with 1b783e53 for Wizlab entries, move Vaults
branch Desolation entries into rooms. Add floor border to some entries
so they work better as Vaults rooms. This border won't cause any issues
for placement outside of Vaults (i.e. in Elf), so I'm not going to CLEAR
it outside of Vaults branch. One entry is a large diamond shape that
won't work terribly well as a Vaults room, so it has 0 weight only for
Vaults.