Make Corpse Rot create more clouds with higher power
Corpse Rot was one of the few player-castable spells that ignored spellpower
completely, and was also fairly awkward to make good use of despite the
strength of miasma clouds. Allow it to rot multiple corpses from a single
stack, and give it a power-dependent chance of creating an extra miasma
cloud per corpse in LOS. With high power and/or high numbers of corpses,
also allow it to spread out to radius 2, once all adjacent squares have
been filled with miasma.