Adjust targeter_radius behaviour
Change a number of radius spell targeters so that the caster's tile isn't
automatically displayed as affected, unless it's actually a spell that can
affect the caster such as Static Discharge. For Battlesphere, also slightly
adjust the spell's behaviour so that the valid squares for recalling an
existing battlesphere are the same as the squares for summoning a new one.