Some fixes and adjustments to some float vaults
For `dreamdust_black_sun_cult`, use the new correct name for demonspawn
black suns. Tweak placement to allow on branch ends, since there's
nothing special about this vault that needs to avoid that.

For `dreamdust_experimentation`, mix in some "experimental" hostile
monsters along with the wizards, to make the vault a bit less mundane.
Scale loot and monster count better with depth of placement.  For the
decorative monsters, use either only human or animal species monsters
depending on the monster set chosen, all with permanent enchants, and
have them generate awake.

For `dreamdust_cat_pyramid`, slightly increase the number of sphinxes,
placing 3 on-average and up to 5 to balance the vault's difficulty
relative to its loot. Use the good loot potion set instead of random
potions, and thin out the loot count a little. This way there's a bit
less loot, but it's higher quality. Spread the loot items out in the
vaults interior, moving external gold to inside the vault. Tweak the
vault layout to prevent statues from blocking the door and to give a
more complete path around the outside of the vault.