Make Passage of Golubria's range scale with power
Passage of Golubria was one of the few remaining spells that ignored
spellpower completely, and was a desirable tool for anyone able to cast it.
To make it less useful with low investment, scale its range with
spellpower, from 2 at 0 power up to full LOS at 100 power. For practical
spellpower values, this results in range 4/5 with minimal skill to cast the
spell, and range 6/7 with more skill investment.