Upgrade Corpse Rot and PoG targeters
Correctly handle corridors and adjacent walls for Corpse Rot, to show when
it has a chance of spilling out and when it's guaranteed to affect a cell.
Highlight squares that overlap for both portals in the Passage of Golubria
targeter, to make it visually clearer that they're two distinct target
areas.