feat: Some tweaks to a Shoals vault
For `dreamdust_from_the_deep`, the custom water and wall tile visuals
taken from the Sewer portal are out of place for the visuals of Shoals.
The vault's themes directly referencing the Abyss, so this commit
removes the custom water tiles and replaces the Sewer wall tiles with
Abyss rock wall tiles.

Also replace the gravestone tile for the statue with the newly added
tentacle statue tile and remove the custom description. Custom statue
descriptions used only in one vault tend to confuse the player as to
whether the statue might some special effect, which this one doesn't.