docs: MSYS2 install tweaks for python issues
The main contentful change here is to be consistent about using the
binary name `python3` instead of `python`. The reason for this is that
apparently these days `base-devel` forces install of the `python`
package as a dep of various things included in it, and we also install
the `mingw-w64-x86_64-python` package as part of that toolchain. Having
the latter installed (which uses the `python3` binary name) seems to
break regular `python` in current MSYS2 versions, because when you load
`python` the python path (as seen in `sys.path`) gets set wrong, to the
mingw toolchain version. Because of this it fails to load any c libs due
to binary incompatibility (`math` being the one that shows up when
trying to import `yaml`).

Anyways, this confusing situation doesn't actually impact building
crawl, because the build steps themselves already use `python3`, it just
means this step that was supposed to check `yaml` was producing an
error. The situation is possibly an outright MSYS2 bug, though I'm not
entirely sure where to report or ask about it. Because the build
actually works fine as-is, I don't think any deeper change is merited on
our part (e.g. uninstalling one or the other python, each of which is a
dep to a corresponding version of gdb).

This commit resolves #2195, which proposed a less direct workaround for
getting an attempt to import `yaml` not to crash.