Make the player immune to their own Noxious Bog
The spell's utility was previously very heavily dependent on whether the
player had easy access to flight. Make it a little easier to use without
needing access to flight, since it already does have strong restrictions
on where it can be cast in the first place. (This can also technically be
used to gain immunity to monster-cast sludge, which is probably fine.)