feat: Some adjustments to two large themed maps
For `dreamdust_labyrinth`, this is a large vault that ultimately places
a single monster. Although the larger maze is visually interesting,
there's a lot of dead space. This commit scales down the maze by
removing one set of stone walls. Additionally it simplifies the
substitutions and introduces an alternating pattern to prevent the dead
end from sometimes placing adjacent to the single opening in the wall in
a way that prevents the dead end from functioning.

For `dreamdust_isle_of_ice`, I've rebalanced the monster lists for each
placement depth and added loot appropriate for the threat. For monsters,
since this vault is evoking the Ice Cave aesthetics and monster themes,
I've made monster lists that combine the "natural" and "undead/demonic"
sets, choosing depth-appropriate monsters from the Ice Cave common
monster lists used in that portal.

The earlier Dungeon and Lair placements use the "easy" Ice Cave
monsters, the later Dungeon uses the "hard" Ice Cave monsters, and for
Depths I've logically extended the sets to include top end "hard"
monsters and a few additional threats form Cocytus. I've added a mix of
the usual loot types as well as one thematic Ice Cave loot item that
borrows some of the loot definitions used in the portal. The monster and
loot item counts are the same for all depths, only the types of monsters
change. Ice statues have a chance to place in Dungeon and Lair, with a
chance for a second ice statue in late Dungeon. For Depths, we use an
orange crystal statue.

I've tweaked the appearance of the vault to be more consistent with the
Ice Cave portal for console, and added vault connectors near the front
of the map, so the vault will be better integrated with the level
layout.