feat: Remove a Gauntlet exit vault
The map `gauntlet_exit_spiral` is a simple spiral Gauntlet exit map
adapted from an exit vault for the original Labyrinth portal. This map
is significantly harder than the other exits because it provides no
cover over long corridors. It predates the minotaur having access to
javelins, so this difficulty difference is unintentional. Furthermore
it's not an especially interesting exit map on its own. The newly added
`dreamdust_gauntlet_exit_labyrinth` map is more balanced and interesting
than `gauntlet_exit_spiral`, and the former is also a Labyrinth
reference, so I'm removing the latter.  It's possible to rework the
map's layout to give the player some better terrain while retaining a
spiral aspect, but it would make more sense to make a new map at that
point.