Cap Gozag's gold distraction chance
In edge cases (primarily casting level 9 spells in Ziggurats), it was
possible to stack gold distraction so high that monsters could barely act.
Adjust the probability of gaining a stack, so that it's now easier to gain
multiple stacks at low values, but harder once lots of stacks have been
gained, and cap the number of stacks at 20. This is unlikely to be relevant
most of the time but should at least tone it down in extreme situtations.