fix: explicitly use TargetConditionals.h in platform detection
OS X provides a header that is in some ways quite similar to our
platform.h, up to the macro naming style, but unfortunately uses a value
of 0 rather than undefinedness for false. This has been around for a
very long time (e.g.  I found a version as old as 2000), and it looks
like there have always been some collisions, but these are largely for
for macos 12) added a few new ones that do matter quite a bit, in
particular TARGET_OS_WINDOWS and TARGET_OS_LINUX are now #defined to 0
when building on OS X. This breaks many things.

This commit takes the easy way out and does some selective #undefs. I've
tried to fix a few of the old problems that don't really matter as well,
e.g. the old version had noticed that TARGET_CPU_PPC was defined but
didn't handle that correctly.  I think in the longer run it could be
better to change our prefixes, e.g. to DCSS_TARGET_OS_WINDOWS. But this
is a bit of an annoying task.

Resolves #2208