Increase the piety cost of Rising Flame
Cinder Acolyte has proven to be an incredibly strong start. In
particular, the Rising Flame power changes the game in ways that
make the entire early and mid-game into a foregone conclusion for
any experienced reasonably-high-win-rate player. While strong
starts such as Berserker already exist, these are all still
susceptible to losing if one makes the various mistakes that
players make in an average game of Crawl. Rising Flame, however,
removes all consequences for mistakes.

At starting Ignis Piety, the player gets four uses of Rising
Flame. When they activate it, they are shafted upwards after a
delay of 2+random2(3) turns (an entire turn faster than a
teleport scroll!). This ability is probably the most powerful
escape ability in the entire game and is simply too strong to be
available from the start of the game in such a large quantity. A
"get out of jail free card" for a player's worst mistakes is a
cool idea, but it should only be available one time, similar to
Tso or Kiku's capstone abilities.

As such, this PR increases the piety cost of Rising Flame by four
times.