fix: make &P work in temple
Previously, the selected map for this command was ignored in temple,
overridden by the temple map chosen at seed start. This commit lets a
FORCE_MAP_KEY prop override TEMPLE_MAP_KEY if both are set. This does
not replace the TEMPLE_MAP_KEY prop, so `&ctrl-r` resets. A map that
requires a layout will crash here; this could maybe be headed off by
requiring the map to be usable in temple to place there in the wizmode
code. I've fixed the crash so that it's an ASSERT at least, and not a
segfault (which is how this initially presented).

This commit also fixes the encompass warning for `&P` to check PLACE
restrictions and not just DEPTH; it was ignoring PLACE before and so
probably printing this warning for a bunch of unnecessary cases.