refactor: do lobby updates on a schedule
Previously, every time any game changed state in a way that would affect
the lobby, the lobby update function was called immediately. This, so
far as I can assess this sort of thing, scales quite poorly with the
number of active players and the number of lobby viewers. The amount of
data was still quite small, but the amount of send_message calls, it
seems, was not. This commit refactors lobby updates to be sent all at
once, by default every 2 seconds. The idea is put a configurable bound
on the amount of traffic involved in lobby info, where lower-traffic
servers could set this much lower (e.g. 1 or 0.5), and higher traffic
servers could tweak upwards as needed.

It should be mentioned that truly profiling webtiles in live situations
is basically impossible as it stands, so this commit is based on some
level of guesswork. It's also not tested yet in a real live situation.