Refactor actor::cannot_move into actor::cannot_act
With the change in 7190d3f, player::cannot_move and player::cannot_act
became equivalent. This was in addition to the monster versions of those
methods also being equivalent, which meant we can remove the super
versions of them.

This also corrects one instance where the lack of checking for
sleepiness in player::cannot_speak was relied on to enable talking while
sleeping.