Improve the monsters in some overflow vaults
Overflow vaults with a uniq_altar tag can place in Depths. Some of these vaults
place very easy monsters, like goblins with spears, which are suitable for the
overflow altar range but not suitable later. This commit either adds an
alternative monster set for Depths placement or limits the range of the vault
to dungeon only, depending on which method seemed best.