feat: transmit milestones/where info over webtiles sockets
In order to update the lobby, the webtiles server needs various bits of
information about ongoing games.  Previously, this information was
transmitted via 0-2 distinct channels: the .where file in the morgue
directory, and the milestone tailing. (The 0 is because on just a plain
`WEBTILES=y` build, where/milestone info was just extremely unreliable,
and milestones weren't sent at all as far as I can tell.)

This commit refactors things so that where info and milestones can just
be sent directly to the webtiles server over the socket connect, rather
than increasing the I/O load. (Probably not impactful on most servers,
but possibly actually matters on CAO.) This change also simplifies
client/server communication quite a bit. For versions with this patch,
the webtiles server does not need to tail the milestones file, and does
not need to read .where files at all, simplifying configuration as well.
We do keep these capabilities around for legacy reasons, as I expect to
backport this a bit but not very far back. Specific notes:

* The crawl process still writes a .where file if DGL_WHEREIS is
  defined. This is necessary for dgl setups, but probably could be
  dispensed with otherwise.
* If a crawl process sends direct where/milestone info, the server will
  stop paying attention to the indirect versions. This should make using
  older versions fairly seamless.
* On the webtiles server side, some light refactoring so that milestones
  and wheres use the same code, moreso than before at least.
* The last milestone is saved in a prop so that it can be shown when
  reloading, and explicitly synced on game start. (Previously it was
  just a static variable that was not saved.)
* An explicit where is sent directly during character generation.