Remove Invisibility as a player spell
Invisibility is a very powerful effect, and has multiple sources (spell,
potion and scarf ego). The spell is by far the most easily repeatable of
these, making it much stronger and less interesting to make effective use
of. The scarf still remains as a non-consumable source of invisibility, but
is better balanced by the additional drain cost.